+++ RELEASE-NOTES.md
@@ -21,6 +21,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 
 ###  Added
 - `sidre::View` holding array data may now be re-shaped.  See `sidre::View::reshapeArray`.
+- Sina C++ library is now a component of Axom
 - Adds optional dependency on [Open CASCADE](https://dev.opencascade.org). The initial intention is 
 to use Open CASCADE's file I/O capabilities in support of Quest applications.